+++ tests/testthat/test-c.R
@@ -439,11 +439,7 @@ test_that("named empty vectors force named output (#1263)", {
   expect_named(list_unchop(list(x, x), list(int(), int())), chr())
   expect_named(list_unchop(list(x, 1L), list(int(), 1)), "")
 
-  # FIXME: `vec_cast_common()` dropped names
-  # https://github.com/r-lib/vctrs/issues/623
-  expect_failure(
-    expect_named(list_unchop(list(x, 1), list(int(), 1)), "")
-  )
+  expect_named(list_unchop(list(x, 1), list(int(), 1)), "")
 })
 
 # Golden tests -------------------------------------------------------